requirements.txt
g4f==0.5.2.9
moviepy==2.1.2
mutagen==1.47.0
numpy==2.3.3
openai_whisper==20240930
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.3.0
protobuf==6.32.1
Requests==2.32.5
token_count==0.2.1
